+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix(`TraceData`): Do not throw exception on missing data.
 * fix(`MainTransformer`): Improve error reporting on circular references.
+* fix(`ObjectCache`): Change backing store to `WeakMap`.
 
 ## 0.6.6